pyproject.toml
[project]
name = "eventanalyzer"
version = "0.1.0"
description = "Telegram bot for personalized news analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "exa-py>=1.0.0",
    "instructor>=1.0.0",
    "openai>=1.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0.0",
    "ipykernel[dev]>=6.29.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
